El agregar NOTAS no actualiza aun
