+++ lib/screens/screens.dart
@@ -8,6 +8,6 @@ export 'package:entre_tiempos/screens/progress/progress_screen.dart';
 export 'package:entre_tiempos/screens/completed_tasks_screen.dart';
 export 'package:entre_tiempos/screens/concentration_screen.dart';
 export 'package:entre_tiempos/screens/development_screen.dart';
-export 'package:entre_tiempos/screens/tareas7_8_notas.dart';
+export 'package:entre_tiempos/screens/tareas7_8notas/tareas7_8_notas.dart';
 export 'package:entre_tiempos/screens/present_screen.dart';
